ui for profile page created
